(ASC-219) Manage test exit failures with trap
This commit alters the `execute_tests.sh` script to manage failures
with a custom `trap`. This enables all test steps of the script to get
executed, but will return the exit code of the first failing command.

The setup portion of the script is set to fail immediately, while the
test execution will iteratively complete even when one of the suites
fails. The failing exit code of any process is captured in a global `RC`
variable and is returned as the scripts exit code. This is the exit code
that is returned to the calling script used by Jenkins.

This ensures the Jenkins job will be marked as failed if a test suite
fails while enabling all tests to complete their execution.